iio: adc: at91: don't use the last converted data register
If touchscreen mode is enabled and a conversion request is asked on a
channel, the result in the last converted data register can be a
touchscreen relative value. Starting a conversion involve to do a
conversion for all active channel. It starts with ADC channels and ends
with touchscreen channels. Then if ADC_LCD register is not read quickly,
its content may be a touchscreen conversion. So it's better to use the
channel data registers.